requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
lxml==5.3.1
packaging==24.2
psycopg==3.2.4
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyPDF2==3.0.1
python-docx==1.1.2
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
